modified delete photo

+++ admin/delete_photo.php
@@ -1,13 +1,10 @@
 <?php include("includes/init.php"); ?>
+<?php if (!$session->is_signed_in()) redirect("login_page.php"); ?>
 
 <?php
-	if (!$session->is_signed_in()) redirect("login_page.php");
-
-$photo = Photo::get_by_id($_GET['id']);
-$photo->delete();
-redirect("admin/photos.php");
-
-
-
-
+	empty($_GET['id']) ?? redirect('admin/photos.php');
+	if ($photo = Photo::get_by_id($_GET['id'])) {
+		$photo->delete();
+}
+	redirect("admin/photos.php");
 ?>
